Added bash script to help manage the way emacs is run
The ecukes command is now a bash script.
ecukes was renamed to ecukes.el, which is now loaded by the script.
You may run ecukes in a graphical way by running it with the argument --graphical
That should clearly be renamed, but I am currently unimaginative.
You may also specify the argument that ecukes uses to run emacs via
the shell variable ECUKES_EMACS; set it to the path of the emacs you
would like it to run.