Setup directory paths relative to current buffer each time ecukes is run
When the ecukes shell script is run from project directory, defvar works
fine. But when run from inside emacs, ecukes is loaded on startup and was
setting the paths relative to where emacs was started. So, unless careful
to start emacs from project directory, the paths were not correct.

By changing defvars to defuns in ecukes-startup, the paths are calculated
each time ecukesmis run. Paths are then are then set relative to the
current buffer and this allows ecukes to be run inside emacs.