Basic support for scenario outlines
The primary shortcoming of this implementation is that
it generates scenarios using the outlines at parse-time,
which results in the ecukes output being the scenarios
rather than the outlines that cucumber proper would
print out.

But if you really want scenario outlines, like I did,
this is better than nothing!