Make verbose default.
Emacs byte compilation uses `princ', which is adviced in
ecukes-core.el. Since the message is only printed if `ecukes-verbose'
and `ecukes-message' is true, which they are not by default, some byte
compilation characters were not printed.

By setting `ecukes-message' to true by default and adding a --quiet
option this is hopefully solved.