DWIM under --debug; Test the working tree, not latest MELPA
The debugger is not reentrant, and invoking --debug at cli uselessly
produced a backtrace for the condition-case call in
`custom-initialize-reset` (because debug-on-signal is true).

Turn off debug-on-signal, and use `condition-case-unless-debug` in
`ecukes-run-step` to get a useful backtrace on the first substantive
error under `--debug`.

Use `cask link` to test the working tree, and not the latest tree on
melpa (as gotten by `cask install`).

Closes #175.